pcf8563: Initial version
This adds an initial version of the PCF8563 real-time clock module. The
code is a loose port of https://github.com/tuupola/pcf8563 which is
written in C.

The M5StickCPlus contains the BM8563, which is a clone of the PCF8563
with a compatible I2C interface. The BM8563 does not handle the turn of
the century correctly.

This initial does not yet have support for timers or alarms.